ENH: Add raise_status option
Add a new option to all functions called ``raise_status``. If ``False``
no exception will be raised and instead ``None`` is returned for those
requests that led to exceptions. This will allow for returning all responses
that were successful and ignoring the ones that failed. This option defaults
to ``True`` for retaining backward compatibility.